refactor(utils): rename isURL to inspectURL
test(inspectURL): add test for isPlaylistURL
